+++ PhysPix Sandbox/src/game/cell/Lava.cpp
@@ -12,7 +12,7 @@
 
 namespace cell
 {
-	Lava::Lava(CellGrid* cellGrid, int32_t posX, int32_t posY)
+	Lava::Lava(CellGrid* cellGrid, uint16_t posX, uint16_t posY)
 	{
 		m_CellGrid = cellGrid;
