+++ PhysPix Sandbox/src/game/cell/Smoke.h
@@ -8,7 +8,7 @@ namespace cell
 	class Smoke : public Cell, public Fluid
 	{
 	public:
-		Smoke(CellGrid* cellGrid, int32_t posX, int32_t posY);
+		Smoke(CellGrid* cellGrid, uint16_t posX, uint16_t posY);
 		~Smoke();
 
 		void OnTick() override;
