+++ PhysPix Sandbox/src/game/cell/Water.cpp
@@ -8,7 +8,7 @@
 
 namespace cell
 {
-	Water::Water(CellGrid* cellGrid, int32_t posX, int32_t posY)
+	Water::Water(CellGrid* cellGrid, uint16_t posX, uint16_t posY)
 	{
 		m_CellGrid = cellGrid;
