+++ PhysPix Sandbox/src/game/cell/Water.h
@@ -8,7 +8,7 @@ namespace cell
 	class Water : public Cell, public Fluid
 	{
 	public:
-		Water(CellGrid* cellGrid, int32_t posX, int32_t posY);
+		Water(CellGrid* cellGrid, uint16_t posX, uint16_t posY);
 		~Water();
 
 		void OnTick() override;
